move circuit_stimulus_definitions into stimulus
