+++ bluecellulab/stimulus/circuit_stimulus_definitions.py
@@ -11,7 +11,11 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+"""Defines the expected data structures associated with the stimulus defined in
+simulation configs.
 
+Run-time validates the data via Pydantic.
+"""
 from __future__ import annotations
 
 from enum import Enum
